perf: reduce token usage and enforce execution plan stage skipping
Token reduction:
- Rebuild agent fresh per stage instead of reusing one instance
  across the full workflow — eliminates accumulated conversation
  history (was the primary cause of 5.38M input tokens / $17 run)
- Reduce SlidingWindowConversationManager window_size 40 → 10
- Simplify token collection to use TokenCountingHook directly
  (agent instances are now ephemeral)

Adaptive workflow:
- Add _get_skipped_stages() — parses execution-plan.md after
  workflow-planning completes and extracts stages marked SKIP
  (strips markdown bold markers and HTML tags before matching)
- Wire plan_skips into both inception and construction stage loops;
  skipped stages print "⏭ (plan: SKIP)" without invoking the agent
- For the "update my profile" story: 6 of 13 stages now correctly
  skipped (application-design, units-generation, functional-design,
  nfr-requirements, nfr-design, infrastructure-design)

+++ ai-dlc-agent/agentcore/.cache/aidlcagent/staging/strands/models/ollama.py
@@ -280,7 +280,7 @@ class OllamaModel(Model):
                             "totalTokens": event["data"].eval_count + event["data"].prompt_eval_count,
                         },
                         "metrics": {
-                            "latencyMs": event["data"].total_duration / 1e6,
+                            "latencyMs": int(event["data"].total_duration / 1e6),
                         },
                     },
                 }
